# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SS Milhas API",
    "version": "1.0.0",
    "description": "API completa do Sistema de Gestão de Milhas Aéreas",
    "contact": {
      "name": "Suporte SS Milhas",
      "email": "suporte@ssmilhas.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Servidor de Desenvolvimento"
    },
    {
      "url": "https://ss-milhas.vercel.app/api",
      "description": "Servidor de Produção"
    }
  ],
...
  "paths": {
...
    "/movements": {
      "get": {
        "tags": ["Movimentações"],
        "summary": "Listar movimentações",
        "description": "Retorna lista de movimentações com filtros e paginação",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tipo",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["compra", "venda", "transferencia", "bonificacao", "expiracao", "estorno"]
            }
          },
          {
            "name": "programa",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pendente", "confirmado", "cancelado"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de movimentações"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Movimentações"],
        "summary": "Criar movimentação",
        "description": "Cria uma nova movimentação de milhas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovementInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Movimentação criada com sucesso"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/tasks": {
      "get": {
        "tags": ["Tarefas"],
        "summary": "Listar tarefas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tarefas"
          }
        }
      },
      "post": {
        "tags": ["Tarefas"],
        "summary": "Criar tarefa",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarefa criada com sucesso"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "MovementInput": {
        "type": "object",
        "required": ["tipo", "programa", "quantidade", "cpf"],
        "properties": {
          "tipo": {
            "type": "string",
            "enum": ["compra", "venda", "transferencia", "bonificacao", "expiracao", "estorno"]
          },
          "programa": {
            "type": "string"
          },
          "quantidade": {
            "type": "integer",
            "minimum": 1
          },
          "valorUnitario": {
            "type": "number",
            "minimum": 0
          },
          "valorTotal": {
            "type": "number",
            "minimum": 0
          },
          "cpf": {
            "type": "string",
            "pattern": "^[0-9]{11}$"
          },
          "descricao": {
            "type": "string"
          }
        }
      },
...
      "TaskInput": {
        "type": "object",
        "required": ["titulo"],
        "properties": {
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["todo", "in_progress", "review", "done"]
          },
          "prioridade": {
            "type": "string",
            "enum": ["baixa", "media", "alta", "urgente"]
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "default": false
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Requisição inválida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Não autorizado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}